# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Main"
        ],
        "summary": "Index",
        "operationId": "index__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/student/lessons": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Get Lessons",
        "operationId": "get_lessons_student_lessons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lessons"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/student/info": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Get Student Info",
        "operationId": "get_student_info_student_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/student/gpa": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Get Student Percentile",
        "operationId": "get_student_percentile_student_gpa_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnExHistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/groups": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Get Groups",
        "operationId": "get_groups_teacher_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/mark": {
      "post": {
        "tags": [
          "Teacher"
        ],
        "summary": "Add Mark",
        "operationId": "add_mark_teacher_mark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstForStudent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/student/{student_code}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Get Student",
        "operationId": "get_student_teacher_student__student_code__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "student_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teacher/lessons": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Teacher Lessons",
        "operationId": "teacher_lessons_teacher_lessons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonsTeacher"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/marking/students": {
      "post": {
        "tags": [
          "Teacher"
        ],
        "summary": "Marking Students",
        "operationId": "marking_students_teacher_marking_students_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentsOnLesson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/disciplines": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Get Disciplines",
        "operationId": "get_disciplines_teacher_disciplines_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Disciplines"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teacher/group/{group_code}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Get Group",
        "operationId": "get_group_teacher_group__group_code__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "group_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/neutral/lesson/{lesson_id}": {
      "get": {
        "tags": [
          "Neutral"
        ],
        "summary": "Get Lesson",
        "operationId": "get_lesson_neutral_lesson__lesson_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/neutral/enter/exit": {
      "post": {
        "tags": [
          "Neutral"
        ],
        "summary": "Enter Exit",
        "description": "\u0422\u0443\u0442 \u043d\u0430\u0434\u043e \u043f\u0435\u0440\u0435\u0434\u0430\u0432\u0430\u0442\u044c \u0442\u043e\u043a\u0435\u043d \u0434\u043e\u0441\u0442\u0443\u043f\u0430, enter_token \u0432\u044b\u0434\u0430\u044e\u0449\u0438\u0439\u0441\u044f admin",
        "operationId": "enter_exit_neutral_enter_exit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEnterExit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/neutral/entrances": {
      "get": {
        "tags": [
          "Neutral"
        ],
        "summary": "Get Entrances",
        "operationId": "get_entrances_neutral_entrances_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnExHistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/neutral/percentile": {
      "get": {
        "tags": [
          "Neutral"
        ],
        "summary": "Get Gpa",
        "operationId": "get_gpa_neutral_percentile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnExHistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/neutral/enter/token": {
      "get": {
        "tags": [
          "Neutral"
        ],
        "summary": "Get Enter Token End",
        "operationId": "get_enter_token_end_neutral_enter_token_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddEnterExit": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "branch_id": {
            "type": "integer",
            "title": "Branch Id"
          }
        },
        "type": "object",
        "required": [
          "time",
          "status",
          "branch_id"
        ],
        "title": "AddEnterExit"
      },
      "Disciplines": {
        "properties": {
          "lessons_and_related_groups": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Lessons And Related Groups"
          }
        },
        "type": "object",
        "required": [
          "lessons_and_related_groups"
        ],
        "title": "Disciplines"
      },
      "EnEx": {
        "properties": {
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "branch_address": {
            "type": "string",
            "title": "Branch Address"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "branch_name",
          "branch_address",
          "time",
          "status"
        ],
        "title": "EnEx"
      },
      "EnExHistory": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/EnEx"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "history"
        ],
        "title": "EnExHistory"
      },
      "EstForStudent": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "mark": {
            "type": "integer",
            "title": "Mark"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "lesson_id",
          "mark"
        ],
        "title": "EstForStudent"
      },
      "GroupInfo": {
        "properties": {
          "groups": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/StudentGroupInfo"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "groups"
        ],
        "title": "GroupInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Lesson": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "auditory_name": {
            "type": "string",
            "title": "Auditory Name"
          },
          "auditory_capacity": {
            "type": "integer",
            "title": "Auditory Capacity"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "branch_address": {
            "type": "string",
            "title": "Branch Address"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "teacher_name": {
            "type": "string",
            "title": "Teacher Name"
          },
          "teacher_secondname": {
            "type": "string",
            "title": "Teacher Secondname"
          },
          "teacher_lastname": {
            "type": "string",
            "title": "Teacher Lastname"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline"
          },
          "type_of_lesson": {
            "type": "string",
            "title": "Type Of Lesson"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "group",
          "auditory_name",
          "auditory_capacity",
          "branch_name",
          "branch_address",
          "start_time",
          "end_time",
          "teacher_name",
          "teacher_secondname",
          "teacher_lastname",
          "task",
          "deadline",
          "type_of_lesson"
        ],
        "title": "Lesson"
      },
      "LessonTeacher": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "auditory_name": {
            "type": "string",
            "title": "Auditory Name"
          },
          "auditory_capacity": {
            "type": "integer",
            "title": "Auditory Capacity"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "branch_address": {
            "type": "string",
            "title": "Branch Address"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline"
          },
          "type_of_lesson": {
            "type": "string",
            "title": "Type Of Lesson"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "group",
          "auditory_name",
          "auditory_capacity",
          "branch_name",
          "branch_address",
          "start_time",
          "end_time",
          "task",
          "deadline",
          "type_of_lesson"
        ],
        "title": "LessonTeacher"
      },
      "Lessons": {
        "properties": {
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/Lesson"
            },
            "type": "array",
            "title": "Lessons"
          }
        },
        "type": "object",
        "required": [
          "lessons"
        ],
        "title": "Lessons"
      },
      "LessonsTeacher": {
        "properties": {
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonTeacher"
            },
            "type": "array",
            "title": "Lessons"
          }
        },
        "type": "object",
        "required": [
          "lessons"
        ],
        "title": "LessonsTeacher"
      },
      "Student": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "secondname": {
            "type": "string",
            "title": "Secondname"
          },
          "lastname": {
            "type": "string",
            "title": "Lastname"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "institute": {
            "type": "string",
            "title": "Institute"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "secondname",
          "lastname",
          "group",
          "institute"
        ],
        "title": "Student"
      },
      "StudentGroupInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "secondname": {
            "type": "string",
            "title": "Secondname"
          },
          "lastname": {
            "type": "string",
            "title": "Lastname"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "secondname",
          "lastname",
          "code"
        ],
        "title": "StudentGroupInfo"
      },
      "StudentsOnLesson": {
        "properties": {
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "students": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Students"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "students"
        ],
        "title": "StudentsOnLesson"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}